requirements/base.txt
Django==1.9.7
django-cors-headers==1.1.0
djangorestframework==3.3.3
PyYAML==3.11
